+++ diagnosticism/__init__.py
@@ -9,7 +9,7 @@ __email__       =   'matthew@synesis.com.au'
 __license__     =   'BSD-3-Clause'
 __maintainer__  =   'Matt Wilson'
 __status__      =   'Beta'
-__version__     =   '0.12.1'
+__version__     =   '0.13.0'
 
 from .contingent_reporting import (
     abort,
@@ -41,4 +41,7 @@ from .tracing import (
     trace,
 )
 from .warning import warn
+
+
+# ############################## end of file ############################# #