+++ diagnosticism/program_name.py
@@ -44,3 +44,6 @@ def get_program_name():
 # initialise the program name
 set_program_name(None)
 
+
+# ############################## end of file ############################# #
+