+++ tests/test_conrep.py
@@ -3,6 +3,7 @@
 from diagnosticism.contingent_reporting import (
     abort,
     conrep,
+    report,
     set_default_trailing_prompt,
 )
 from diagnosticism.program_name import set_program_name
@@ -29,7 +30,7 @@ class ConRep_tester(unittest.TestCase):
 
             conrep('string-1', show_program_name=True)
 
-            self.assertEqual('myprog1: string-1\n', fake_stderr.getvalue())
+            self.assertEqual("myprog1: string-1\n", fake_stderr.getvalue())
 
 
     def test_without_program_name(self):
@@ -38,7 +39,56 @@ class ConRep_tester(unittest.TestCase):
 
             conrep('string-1', show_program_name=False)
 
-            self.assertEqual('string-1\n', fake_stderr.getvalue())
+            self.assertEqual("string-1\n", fake_stderr.getvalue())
+
+
+class Report_tester(unittest.TestCase):
+
+
+    def test_with_program_name(self):
+
+        with patch('sys.stderr', new=StringIO()) as fake_stderr:
+
+            set_program_name('myprog1')
+
+            report('string-1', show_program_name=True)
+
+            self.assertEqual("myprog1: string-1\n", fake_stderr.getvalue())
+
+
+    def test_without_program_name(self):
+
+        with patch('sys.stderr', new=StringIO()) as fake_stderr:
+
+            report('string-1', show_program_name=False)
+
+            self.assertEqual("string-1\n", fake_stderr.getvalue())
+
+
+    def test__report__WITH_file_PARAM_AND_show_program_name_False(self):
+
+        file = StringIO()
+
+        set_program_name('myprog3')
+
+        report('string-3', file=file, show_program_name=False)
+
+        result = file.getvalue()
+
+        self.assertEqual("string-3\n", result)
+
+
+    def test__report__WITH_file_PARAM_AND_show_program_name_True(self):
+
+        file = StringIO()
+
+        set_program_name('myprog4')
+
+        report('string-4', file=file, show_program_name=True)
+
+        result = file.getvalue()
+
+        self.assertEqual("myprog4: string-4\n", result)
 
 
 class Abort_tester(unittest.TestCase):
@@ -52,7 +102,7 @@ class Abort_tester(unittest.TestCase):
 
             abort('over and out!', do_exit=False)
 
-            self.assertEqual('myprog1: over and out!\n', fake_stderr.getvalue())
+            self.assertEqual("myprog1: over and out!\n", fake_stderr.getvalue())
 
 
     def test_explicit_trailing_prompt(self):
@@ -63,7 +113,7 @@ class Abort_tester(unittest.TestCase):
 
             abort('over and out!', do_exit=False, trailing_prompt='get over yourself!')
 
-            self.assertEqual('myprog1: over and out!; get over yourself!\n', fake_stderr.getvalue())
+            self.assertEqual("myprog1: over and out!; get over yourself!\n", fake_stderr.getvalue())
 
 
     def test_stock_trailing_prompt(self):
@@ -74,10 +124,10 @@ class Abort_tester(unittest.TestCase):
 
             abort('over and out!', do_exit=False, trailing_prompt=True)
 
-            self.assertEqual('myprog1: over and out!; use --help for usage\n', fake_stderr.getvalue())
+            self.assertEqual("myprog1: over and out!; use --help for usage\n", fake_stderr.getvalue())
 
 
-    def test_set_default_trailing_prompt(self):
+    def test_set_default_trailing_prompt_1(self):
 
         with patch('sys.stderr', new=StringIO()) as fake_stderr:
 
@@ -89,10 +139,10 @@ class Abort_tester(unittest.TestCase):
 
             set_default_trailing_prompt(None)
 
-            self.assertEqual('myprog1: over and out!; ok, now\n', fake_stderr.getvalue())
+            self.assertEqual("myprog1: over and out!; ok, now\n", fake_stderr.getvalue())
 
 
-    def test_set_default_trailing_prompt(self):
+    def test_set_default_trailing_prompt_2(self):
 
         with patch('sys.stderr', new=StringIO()) as fake_stderr:
 
@@ -104,8 +154,7 @@ class Abort_tester(unittest.TestCase):
 
             set_default_trailing_prompt(None)
 
-            self.assertEqual('myprog1: over and out!\n', fake_stderr.getvalue())
-
+            self.assertEqual("myprog1: over and out!\n", fake_stderr.getvalue())
 
 
 if '__main__' == __name__: